app/recepcion/requirements.txt
kafka-python-ng
psycopg2-binary
prometheus-client
